+++ zeb/src/main/java/site/zbyte/zeb/callback/ICallback.kt
@@ -1,5 +1,5 @@
 package site.zbyte.zeb.callback
 
 interface ICallback {
-    fun getId():Long
+    fun getId():Int
 }
